#1415 consider npm config when fetching package list  Adds support for per-project .npmrc files (and multiple registry configurations)  - Splits database so that a separate database corresponds to each registry. That way users can work with different projects/registries and seamlessly switch between .npmrc files using packages like npmrc.  - The database was split into multiple files because didn't want the horrible user experience of assuming a database is corrupt, and deleting all registries cache due to one bad egg. Also makes for quicker db queries in the event a user is using multiple npm registries.  - executes npm get registry command in project folder instead of solution folder  - database schema version bump as this includes breaking changes  - updated tests
 Misc. UI fixes
 - error messages
 - progress bar not appearing when reading database
